create map component
