+++ management/src/modules/events/modalAdd.js
@@ -4,51 +4,7 @@
     import { yupResolver } from '@hookform/resolvers/yup'
     import { BackendClientRequest } from "../../services/ApiClient";
     import { Circle, GoogleMap, Marker, useJsApiLoader } from "@react-google-maps/api";
-
-    const MapContainer = ({ markerPosition, setMarkerPosition, setValue, circleRadius }) => {
-
-        const { isLoaded } = useJsApiLoader({
-            id: "google-map-script",
-            googleMapsApiKey: "AIzaSyAxDq9p3_nTuQ3Vw366xuSV8RfZGlusFvU", // Replace with your own API key
-        });
-    
-        const mapOptions = {
-            center: { lat: 51.74583, lng: 5.63194 }, // Initial center location
-            zoom: 14, // Initial zoom level
-        };
-    
-        const handleMapClick = (event) => {
-            const { lat, lng } = event.latLng;
-            const latValue = lat();
-            const lngValue = lng();
-            setMarkerPosition({ lat: latValue, lng: lngValue });
-            setValue("latitude", latValue.toString());
-            setValue("longitude", lngValue.toString());
-            // setMarkerPosition({ lat: lat(), lng: lng() });
-        };
-
-        
-        return isLoaded ? (
-        <GoogleMap
-            mapContainerStyle={{ height: "300px", width: "100%" }}
-            options={mapOptions}
-            onClick={handleMapClick}
-        >
-            <Marker position={markerPosition}/>
-            <Circle
-                center={markerPosition}
-                radius={circleRadius} // Radius in meters
-                options={{
-                    strokeColor: "blue",
-                    strokeOpacity: 0.8,
-                    strokeWeight: 2,
-                    fillColor: "blue",
-                    fillOpacity: 0.2,
-                }}
-            />
-        </GoogleMap>
-        ) : null;
-    };
+    import MapContainer from "../../components/map";
 
     const ModalAdd = ({ toggleModal, refreshOverview, date, isParade }) => {
         const [showLocationInputs, setShowLocationInputs] = useState(!isParade);
